Move cf-api.js to src/background.
Don't know why I put it in src/common :thonk:
